# server/helpers/swagger.yaml
basePath: /
consumes:
- application/json
- multipart/form-data
definitions:
...
  PerformanceTestConfig:
    properties:
      clients:
        description: Single or distributed load generators
        items:
          $ref: '#/definitions/PerformanceTestConfig_Client'
        type: array
        x-go-name: Clients
      duration:
        description: Length of time the endpoint will be under load
        type: string
        x-go-name: Duration
      id:
        type: string
        x-go-name: Id
      labels:
        additionalProperties:
          type: string
        type: object
        x-go-name: Labels
      name:
        type: string
        x-go-name: Name
      smp_version:
        description: Spec version
        type: string
        x-go-name: SmpVersion
    type: object
    x-go-package: github.com/layer5io/service-mesh-performance/spec
  PerformanceTestConfig_Client:
    properties:
      body:
        type: string
        x-go-name: Body
      connections:
        format: int32
        type: integer
        x-go-name: Connections
      content_type:
        type: string
        x-go-name: ContentType
      cookies:
        additionalProperties:
          type: string
        type: object
        x-go-name: Cookies
      endpoint_urls:
        items:
          type: string
        type: array
        x-go-name: EndpointUrls
      headers:
        additionalProperties:
          type: string
        type: object
        x-go-name: Headers
      internal:
        type: boolean
        x-go-name: Internal
      load_generator:
        type: string
        x-go-name: LoadGenerator
      protocol:
        $ref: '#/definitions/PerformanceTestConfig_Client_Protocol'
      rps:
        format: int64
        type: integer
        x-go-name: Rps
    type: object
    x-go-package: github.com/layer5io/service-mesh-performance/spec
  PerformanceTestConfig_Client_Protocol:
    format: int32
    type: integer
    x-go-package: github.com/layer5io/service-mesh-performance/spec
...
info:
  description: |-
    the purpose of this application is to provide an application
    that is using plain go code to define an API

    This should demonstrate all the possible comment annotations
    that are available to turn go code into a fully compliant swagger 2.0 spec
  license:
    name: Apache-2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.txt
  title: Meshery API.
  version: 0.4.27
paths:
...
  /api/user/prefs/perf:
    delete:
      description: Used for deleting load test preferences
      operationId: idDeleteLoadPreferences
      parameters:
      - format: uuid
        in: query
        name: uuid
        type: string
        x-go-name: UUID
      responses:
        "200":
          description: ""
      summary: Handle DELETE request for load test preferences
      tags:
      - UserAPI
    get:
      description: Used for fetching load test preferences
      operationId: idGetLoadPreferences
      parameters:
      - format: uuid
        in: query
        name: uuid
        type: string
        x-go-name: UUID
      responses:
        "200":
          $ref: '#/responses/loadTestPreferencesWrapper'
      summary: Handle GET request for load test preferences
      tags:
      - UserAPI
    post:
      description: Used for persisting load test preferences
      operationId: idPostLoadPreferences
      parameters:
      - in: body
        name: Body
        schema:
          $ref: '#/definitions/PerformanceTestConfig'
      responses:
        "200":
          description: ""
      summary: Handle POST request for load test preferences
      tags:
      - UserAPI
...
produces:
- application/json
responses:
...
  loadTestPreferencesWrapper:
    description: Returns load test preferences
    schema:
      $ref: '#/definitions/PerformanceTestConfig'
...
schemes:
- http
security:
- token:
  - '[]'
securityDefinitions:
  token:
    in: cookie
    name: token
    type: apiKey
swagger: "2.0"